backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
blinker==1.9.0
blis==1.3.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.2.0
cryptography==46.0.4
cycler==0.12.1
cymem==2.0.13
distlib==0.3.9
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.128.0
filelock==3.16.1
flake8==7.3.0
Flask==3.1.0
fonttools==4.44.3
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-datastore==2.20.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.5
MarkupSafe==3.0.3
matplotlib==3.8.2
mccabe==0.7.0
mongomock==4.3.0
motor==3.7.1
murmurhash==1.0.15
nltk==3.9.2
numpy==1.26.4
packaging==26.0
pandas==3.0.0
passlib==1.7.4
Pillow==10.1.0
platformdirs==4.3.6
pluggy==1.6.0
preshed==3.0.12
proto-plus==1.26.1
protobuf==5.29.4
pyasn1>=0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pymongo==4.16.0
pyparsing==3.1.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
regex==2026.1.15
requests==2.32.5
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
sentinels==1.1.1
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.50.0
thinc==8.3.10
threadpoolctl==3.6.0
tqdm==4.67.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
virtualenv==20.26.6
wasabi==1.1.3
weasel==0.4.3
Werkzeug==3.1.3
wrapt==2.0.1
